pyproject.toml
[project]
name = "runreview"
version = "0.1.0"
description = "AI-powered codebase analysis platform"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "gitpython",
    "groq",
    "python-dotenv",
    "pydantic-settings",
    "litellm",
    "PyYAML",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "httpx",
]
